moved getColorByFullness  to helper, and getOptionLabelWithAvailable to selectUtils
